New changes for pgr

+++ rainmaker-pgr/src/main/java/org/egov/pgr/model/GrievenceReport.java
@@ -1,5 +1,6 @@
 package org.egov.pgr.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -38,7 +39,7 @@ public class GrievenceReport {
 	public String slaAchievement;
 
 	@JsonProperty("completionRate")
-	public Double completionRate;
+	public BigDecimal completionRate;
 
 	@JsonProperty("uniqueCitizens")
 	public Integer uniqueCitizens;
